Improve ease of extention
In our usage we need to extend the ControllerInvoker to have responses
normalised from a generic Payload into a standard Response.

To make this easier it would be helpful if some private methods were
instead protected and the Invoker fetching the default resolver chain
were separated out from the creation of the ControllerInvoker.

Fixes #105